requirements-dev.txt
pytest
pytest-django
pytest-cov
flake8
tox
